Correccion al carrito

+++ Art-Shop-Grupo1/Art-Shop-Data/Model/OrderDetail.cs
@@ -12,7 +12,7 @@ namespace Art_Shop_Data.Model
 
         public int ProductId { get; set; }
         
-        public float Price { get; set; }
+        public double Price { get; set; }
 
         public int Quantity { get; set; }
 
@@ -21,7 +21,7 @@ namespace Art_Shop_Data.Model
                 
         }
 
-        public OrderDetail(int id, int OrderId, int ProductId, float Price, int Quantity, DateTime CreatedOn, String CreatedBy, DateTime ChangedOn, String ChangedBy)
+        public OrderDetail(int id, int OrderId, int ProductId, double Price, int Quantity, DateTime CreatedOn, String CreatedBy, DateTime ChangedOn, String ChangedBy)
         {
             this.Id = id;
             this.OrderId = OrderId;
